requirements.txt
pandas
gspread
oauth2client
Pillow
PyGithub